Define CHRE daemon variable and remove rc config
Define CHRE_DAEMON_LOAD_INTO_SENSORSPD and CHRE_DAEMON_ENABLED so that
the CHRE daemon is built for bonito now that it's configured out of system/chre.

Additionally, the daemon's init.rc config is now included as part of the
daemon makefile so remove it from the device's init.hardware.rc.

Bug: 139693714
Test: compile
Change-Id: I35e200cd2d1153b88c9416c27047fa58d9b6daea
Signed-off-by: someone5678 <nemui3353@gmail.com>